The old code would parse the following header field name as the value an empty score header, and length of e.g. "X-Spam-Status:" is 14, a rather high value.
The new regexp requires the spam indicator to consist of just one
character, and that the spam indicator is set off from the rest of
the text.

Since the default Exiscan configuration adds ++++ for positive scores
and ---- for negative scores, we now prefer to look at the numeric
value when one is to be found.